# openapi/core-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/gifts": {
      "get": {
        "operationId": "GiftController_listCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GiftDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Catalog quà đang bật — giá server là nguồn sự thật",
        "tags": ["gift"]
      }
    },
    "/api/v1/party/rooms/{roomId}/gifts": {
      "post": {
        "operationId": "GiftController_sendGift",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendGiftDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/GiftEventDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Tặng quà trong phòng — trừ DIA người tặng + cộng PTS người nhận atomic; client chỉ bắn hiệu ứng sau khi nhận 200",
        "tags": ["gift"]
      }
    },
    "/api/v1/videos/{videoId}/gifts": {
      "post": {
        "operationId": "GiftController_sendVideoGift",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendVideoGiftDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/GiftEventDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Tặng quà cho tác giả video — người nhận suy từ video; cùng chốt tiền/idempotency với quà trong phòng",
        "tags": ["gift"]
      }
    },
...
    "/api/v1/feed/posts": {
      "post": {
        "operationId": "FeedController_createPost",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePostDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PostDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Đăng bài — guest bị chặn (docs/06), audience mặc định public",
        "tags": ["feed"]
      },
      "get": {
        "operationId": "FeedController_listFeed",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PostsPageDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Feed công khai toàn cục, mới nhất trước — loại tác giả đang block/bị block",
        "tags": ["feed"]
      }
    },
    "/api/v1/feed/users/{userId}/posts": {
      "get": {
        "operationId": "FeedController_listUserTimeline",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PostsPageDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Timeline 1 tác giả — audience tự khớp quan hệ (mình/bạn/người lạ), docs/services/feed-service.md § 7",
        "tags": ["feed"]
      }
    },
    "/api/v1/feed/posts/{postId}": {
      "get": {
        "operationId": "FeedController_getPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PostDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Chi tiết 1 bài — 404 nếu đã xoá/bị block (chống oracle)",
        "tags": ["feed"]
      },
      "delete": {
        "operationId": "FeedController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Xoá mềm bài viết — chỉ tác giả",
        "tags": ["feed"]
      }
    },
    "/api/v1/feed/posts/{postId}/comments": {
      "post": {
        "operationId": "FeedController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCommentDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/CommentDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Bình luận — guest bị chặn, chặn nếu block với tác giả",
        "tags": ["feed"]
      },
      "get": {
        "operationId": "FeedController_listComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/CommentsPageDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Danh sách bình luận, cũ → mới",
        "tags": ["feed"]
      }
    },
    "/api/v1/feed/comments/{commentId}": {
      "delete": {
        "operationId": "FeedController_deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Xoá mềm comment — chỉ tác giả",
        "tags": ["feed"]
      }
    },
    "/api/v1/feed/posts/{postId}/reactions": {
      "post": {
        "operationId": "FeedController_like",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReactionStatusDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Thả tim — idempotent, guest bị chặn",
        "tags": ["feed"]
      },
      "delete": {
        "operationId": "FeedController_unlike",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReactionStatusDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Bỏ thả tim — idempotent",
        "tags": ["feed"]
      },
      "get": {
        "operationId": "FeedController_reactionStatus",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReactionStatusDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Trạng thái thả tim của chính mình + tổng đếm",
        "tags": ["feed"]
      }
    },
...
  },
  "info": {
    "title": "Litmatch core-api",
    "description": "",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "GiftDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "priceDiamond": { "type": "number" }
        },
        "required": ["id", "code", "name", "priceDiamond"]
      },
      "SendGiftDto": {
        "type": "object",
        "properties": {
          "giftId": { "type": "string" },
          "receiverUserId": { "type": "string" }
        },
        "required": ["giftId", "receiverUserId"]
      },
      "GiftEventDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "giftId": { "type": "string" },
          "giftCode": { "type": "string" },
          "roomId": { "type": "string", "nullable": true },
          "videoId": { "type": "string", "nullable": true },
          "senderUserId": { "type": "string" },
          "receiverUserId": { "type": "string" },
          "priceDiamond": { "type": "number" },
          "pointsAwarded": {
            "type": "number",
            "description": "0 nếu người nhận là guest (docs/06 § Gift)"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "replayed": { "type": "boolean" }
        },
        "required": [
          "id",
          "giftId",
          "giftCode",
          "senderUserId",
          "receiverUserId",
          "priceDiamond",
          "pointsAwarded",
          "createdAt",
          "replayed"
        ]
      },
      "SendVideoGiftDto": {
        "type": "object",
        "properties": { "giftId": { "type": "string" } },
        "required": ["giftId"]
      },
...
      "ReactionStatusDto": {
        "type": "object",
        "properties": {
          "liked": { "type": "boolean" },
          "likeCount": { "type": "number" }
        },
        "required": ["liked", "likeCount"]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": { "type": "string" },
          "imageUrl": { "type": "string" },
          "audience": {
            "type": "string",
            "enum": ["public", "friends", "only_me"],
            "default": "public"
          }
        }
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "authorUserId": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "imageUrl": { "type": "string", "nullable": true },
          "audience": {
            "type": "string",
            "enum": ["public", "friends", "only_me"]
          },
          "likeCount": { "type": "number" },
          "commentCount": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "authorUserId",
          "content",
          "imageUrl",
          "audience",
          "likeCount",
          "commentCount",
          "createdAt"
        ]
      },
      "PostsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PostDto" }
          },
          "nextCursor": { "type": "string", "nullable": true }
        },
        "required": ["items", "nextCursor"]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": { "content": { "type": "string" } },
        "required": ["content"]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "postId": { "type": "string" },
          "authorUserId": { "type": "string" },
          "content": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "postId", "authorUserId", "content", "createdAt"]
      },
      "CommentsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CommentDto" }
          },
          "nextCursor": { "type": "string", "nullable": true }
        },
        "required": ["items", "nextCursor"]
      },
...
    }
  }
}